% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{as.data.frame.MDFS}
\alias{as.data.frame.MDFS}
\title{as.data.frame S3 method implementation for MDFS}
\usage{
\method{as.data.frame}{MDFS}(x, ...)
}
\arguments{
\item{x}{an MDFS object}

\item{...}{ignored}
}
\value{
data.frame
}
\description{
as.data.frame S3 method implementation for MDFS
}
