% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Discretize}
\alias{Discretize}
\title{Discretize variable on demand}
\usage{
Discretize(data, variable.idx, divisions, discretization.nr, seed, range)
}
\arguments{
\item{data}{input data where columns are variables and rows are observations (all numeric)}

\item{variable.idx}{variable index (as it appears in \code{data})}

\item{divisions}{number of divisions}

\item{discretization.nr}{discretization number (positive integer)}

\item{seed}{seed for PRNG}

\item{range}{discretization range}
}
\value{
Discretized variable.
}
\description{
Discretize variable on demand
}
\examples{
Discretize(madelon$data, 3, 1, 1, 0, 0.5)
}
