% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetRange}
\alias{GetRange}
\title{Get the recommended range for multiple discretisations}
\usage{
GetRange(k = 5, n, dimensions, divisions = 1)
}
\arguments{
\item{k}{the assumed minimum number of objects in a bucket (the default is the recommended value)}

\item{n}{the number of objects in the least numerous class}

\item{dimensions}{the number of dimensions of analysis}

\item{divisions}{the number of divisions of discretisations}
}
\value{
The recommended range value (a floating point number).
}
\description{
Get the recommended range for multiple discretisations
}
\examples{
GetRange(n = 250, dimensions = 2)
}
