% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mdfs_omp_set_num_threads}
\alias{mdfs_omp_set_num_threads}
\title{Call omp_set_num_threads}
\usage{
mdfs_omp_set_num_threads(num_threads)
}
\arguments{
\item{num_threads}{input data where columns are variables and rows are observations (all numeric)}
}
\description{
Call omp_set_num_threads
}
