% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{plot.MDFS}
\alias{plot.MDFS}
\title{Plot MDFS details}
\usage{
\method{plot}{MDFS}(x, plots = c("ig", "c", "p"), ...)
}
\arguments{
\item{x}{an MDFS object}

\item{plots}{plots to plot (ig for max IG, c for chi-squared p-values, p for p-values)}

\item{...}{passed on to \code{\link[graphics]{plot}}}
}
\description{
Plot MDFS details
}
