% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{RelevantVariables.MDFS}
\alias{RelevantVariables.MDFS}
\title{Find indices of relevant variables from MDFS}
\usage{
\method{RelevantVariables}{MDFS}(fs, level = 0.05, p.adjust.method = "holm",
  ...)
}
\arguments{
\item{fs}{an MDFS object}

\item{level}{statistical significance level}

\item{p.adjust.method}{method as accepted by \code{\link[stats]{p.adjust}}}

\item{...}{ignored}
}
\value{
indices of relevant variables
}
\description{
Find indices of relevant variables from MDFS
}
