% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{RelevantVariables}
\alias{RelevantVariables}
\title{Find indices of relevant variables}
\usage{
RelevantVariables(fs, ...)
}
\arguments{
\item{fs}{feature selector}

\item{...}{arguments passed to methods}
}
\value{
indices of important variables
}
\description{
Find indices of relevant variables
}
