% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_value.R
\name{RelevantVariables.MDFS}
\alias{RelevantVariables.MDFS}
\title{Find indices of relevant variables from MDFS}
\usage{
\method{RelevantVariables}{MDFS}(fs, level = 0.05, p.adjust.method = "holm", ...)
}
\arguments{
\item{fs}{an MDFS object}

\item{level}{statistical significance level}

\item{p.adjust.method}{method as accepted by \code{\link[stats]{p.adjust}} (\code{"BY"} is recommended for FDR, see Details)}

\item{...}{ignored}
}
\value{
indices of relevant variables
}
\description{
Find indices of relevant variables from MDFS
}
\details{
In case of FDR control it is recommended to use Benjamini-Hochberg-Yekutieli p-value adjustment
method (\code{"BY"} in \code{\link[stats]{p.adjust}}) due to unknown dependencies between tests.
}
